Reference application data via static `appName` field in `app.js`
Separates user selected data and datastore data into separate stores of
`global` and `datastore`. The `datastore` Redux model is now blacklisted
to support constant updates to server side data.
